+++ package.json
@@ -14,6 +14,7 @@
     "@heroui/react": "^2.7.5",
     "@mui/icons-material": "^7.0.1",
     "@mui/material": "^7.0.1",
+    "@react-oauth/google": "^0.12.1",
     "@reduxjs/toolkit": "^2.6.1",
     "lucide-react": "^0.487.0",
     "react": "^18.0.0",
@@ -23,6 +24,7 @@
     "react-router-dom": "^7.4.0",
     "redux-persist": "^6.0.0",
     "tailwindcss": "^3.0.0",
+    "twglow": "^0.0.10",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
